+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Duplicate responses and timezone issue in quiz calculations [#52](https://github.com/rokmetro/surveys-building-block-fork/issues/52)
 - Issues sending quiz notifications [#59](https://github.com/rokmetro/surveys-building-block-fork/issues/59)
 - Issues sending quiz notifications (bug fixes and optimizations) [#59](https://github.com/rokmetro/surveys-building-block-fork/issues/59#issuecomment-3648478668)
+- Do not force await periodic functions [#62](https://github.com/rokmetro/surveys-building-block-fork/issues/62)
 
 ### Changed
 - Optimize ranking counts [#24](https://github.com/rokmetro/surveys-building-block-fork/issues/24)